// TCOM Client/src/app/core/models/home.ts
export interface Home {
  id: number;
  title: string;
  description: string;
  mainHeader: string;
  serviceHeader: string;
  serviceContent: string;
  techHeader: string;
  techContent: string;
  partnerHeader: string;
  partnerContent: string;
  firstImgUrl: string;
  firstImgId: number;
  secondImgUrl: string;
  secondImgId: number;
  languageId: number;
}